Extract modal keydown helper
